Develop #3 Доделываем

+++ app/controllers/StudentsController.php
@@ -21,25 +21,30 @@ class StudentsController extends BaseController
 
     public function create()
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $data['title'] = 'Создание студента';
         $data['content'] = 'students/create';
-        $data['data'] = Groups::getAll(); # TODO Заменить
+        $data['data'] = (new Groups())->findAll();
         $data['session'] = $this->session;
         return view('includes/template', $data);
     }
 
     public function store()
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $data['name'] = $this->request->getPost('name');
         $data['surname'] = $this->request->getPost('surname');
-        $data['group_id'] =$this->request->getPost('group_id');
+        $data['group_id'] = $this->request->getPost('group_id');
         $obj = new Students();
         try {
             $obj->insert($data);
-        } catch (DatabaseException $e)
-        {
+        } catch (DatabaseException $e) {
             $this->session->set([
-                'msg' => 'Ошибка!'.$e->getMessage(),
+                'msg' => 'Ошибка!' . $e->getMessage(),
                 'msg_type' => 'alert-danger'
             ]);
             return $this->response->redirect(site_url('/students'));
@@ -49,10 +54,12 @@ class StudentsController extends BaseController
 
     public function edit(int $id)
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $obj = new Students();
         $data['data'] = $obj->find($id);
-        if ($data['data'] == null)
-        {
+        if ($data['data'] == null) {
             $this->session->set([
                 'msg' => 'Ошибка! Студент не найден',
                 'msg_type' => 'alert-danger'
@@ -61,13 +68,16 @@ class StudentsController extends BaseController
         }
         $data['title'] = 'Изменение студента';
         $data['content'] = 'students/edit';
-        $data['groups'] = Groups::getAll();
+        $data['groups'] = (new Groups())->findAll();
         $data['session'] = $this->session;
         return view('includes/template', $data);
     }
 
     public function update()
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $id = $this->request->getPost('id');
         $data['name'] = $this->request->getPost('name');
         $data['surname'] = $this->request->getPost('surname');
@@ -75,10 +85,9 @@ class StudentsController extends BaseController
         $obj = new Students();
         try {
             $obj->update($id, $data);
-        } catch (DatabaseException $e)
-        {
+        } catch (DatabaseException $e) {
             $this->session->set([
-                'msg' => 'Ошибка!'.$e->getMessage(),
+                'msg' => 'Ошибка!' . $e->getMessage(),
                 'msg_type' => 'alert-danger'
             ]);
             return $this->response->redirect(site_url('/students'));
@@ -88,8 +97,11 @@ class StudentsController extends BaseController
 
     public function delete(int $id)
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $obj = new Students();
-        $obj->delete($id); # INFO Интересно, что бд не вызывает исключение если удалять несуществующий id
+        $obj->delete($id);
         return $this->response->redirect(site_url('/students'));
     }
 }
